sdk/python/pyproject.toml
[project]
name = "eaasp-sdk"
version = "0.1.0"
description = "EAASP Enterprise SDK — create, validate, and test Skills"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
cli = ["click>=8.0", "rich>=13.0"]
grpc = ["grpcio>=1.60"]
submit = ["httpx>=0.27"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "jsonschema>=4.0"]
all = ["eaasp-sdk[cli,grpc,submit]"]

[project.scripts]
eaasp = "eaasp.cli.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eaasp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
